% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-inquiryTemplate-module.R
\name{inquiryTemplateUI}
\alias{inquiryTemplateUI}
\alias{inquiryTemplateServer}
\title{Inquiry Template UI}
\usage{
inquiryTemplateUI(id)

inquiryTemplateServer(id, inquiry_template)
}
\arguments{
\item{id}{The module id}

\item{inquiry_template}{The inquiry template containing the questions data frame}
}
\description{
Inquiry Template UI

Inquiry Template Server
}
