% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-loadInquiry-module.R
\name{loadInquiryUI}
\alias{loadInquiryUI}
\alias{loadInquiryServer}
\title{Load Inquiry UI}
\usage{
loadInquiryUI(id)

loadInquiryServer(id, submitted_templates)
}
\arguments{
\item{id}{The module id}

\item{submitted_templates}{Reactive value of submitted templates}
}
\description{
Load Inquiry UI

Load Inquiry Server
}
