% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sanitize.R
\name{sanitizeQuestions}
\alias{sanitizeQuestions}
\title{Sanitize Questions}
\usage{
sanitizeQuestions(questions)
}
\arguments{
\item{questions}{The questions to sanitize}
}
\description{
Replace 'dependence' and 'dependence_value' values equal to "" with NA.
}
