% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-sanitize.R
\name{sanitizeQuestionsForJson}
\alias{sanitizeQuestionsForJson}
\title{Sanitize questions for JSON}
\usage{
sanitizeQuestionsForJson(list)
}
\arguments{
\item{list}{A list with a data frame named questions}
}
\value{
A list with NA replaced by list(NULL) in list$questions
}
\description{
Replace NA in list$questions with list(NULL) to avoid empty columns. This prevents toJSON from dropping empty columns.
}
