% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-loadInquiry-module.R
\name{validateImport}
\alias{validateImport}
\title{Validate Import}
\usage{
validateImport(imported_template)
}
\arguments{
\item{imported_template}{The imported template}
}
\description{
Validate Import
}
